Fix Uint8 overflow in the guard-area least-explored direction scan
handleMovementAttackingAround picks the least-known of the eight
neighbouring directions by summing three getExplored() samples per
direction. getExplored returns a Uint8 in [0, 255], so the three-sample
sum reaches 765 and wrapped in the Uint8 accumulator, while the
bestExplored it is compared against is an int seeded at
3*GRADIENT_AT_GOAL (765). A well-explored direction could therefore wrap
to a small value and win the "least known" comparison, sending guarding
units toward terrain they had already explored.

Widen the accumulator to int.

Baselines regenerated: this changes unit movement, so cpp-refactor and
the large-4ai, archipelago and bigarena-long gradient replays no longer
match their old bytes. gd-small-2ai is unaffected.

+++ src/unit/UnitMovement.cpp
@@ -310,7 +310,7 @@ void Unit::handleMovementAttackingAround()
 					if (owner->map->getGuardAreasGradient(posX + cdx, posY + cdy, performance[SWIM]>0, owner->teamNumber) != GRADIENT_AT_GOAL)
 						continue;
 				}
-				Uint8 explored = owner->map->getExplored(posX + 2*cdx, posY + 2*cdy, owner->teamNumber);
+				int explored = owner->map->getExplored(posX + 2*cdx, posY + 2*cdy, owner->teamNumber);
 				explored += owner->map->getExplored(posX + 2*cdx - cdy, posY + 2*cdy + cdx, owner->teamNumber);
 				explored += owner->map->getExplored(posX + 2*cdx + cdy, posY + 2*cdy - cdx, owner->teamNumber);
 				if (bestExplored > explored)
